feat: 控制面升级 PostgreSQL + Redis 缓存；修 alembic 链 + ASN BigInteger
- 控制面支持 PostgreSQL 后端（asyncpg）；ASN 列改 BigInteger（DN42 ASN 超 int32，
  SQLite 动态宽度故历史未暴露，迁 PG 才显形）。
- 加 Redis 缓存层（services/cache.py，best-effort 优雅降级）：缓存 desired-state
  （gen 键不可变）/ fleet·node health（10s）/ routing 聚合（30s），写时主动失效。
- 查询优化：routing churn 的 50k 行 DISTINCT 门控进哈希变更分支；materializer DNS
  记录加载 N+1 改单条 IN 查询。
- 修 alembic 链：初始迁移补建 node_routing/node_routing_events（历史 schema），
  使 alembic upgrade head 从空库可跑通、与 create_all 等价。

+++ apps/control-server/app/db/models/routing.py
@@ -17,7 +17,7 @@ agent 周期上报的 ``RoutingTableSnapshot`` 持久化在这里，与 reconcil
 
 from datetime import datetime, timezone
 
-from sqlalchemy import Boolean, DateTime, ForeignKey, Index, Integer, JSON, String, func
+from sqlalchemy import BigInteger, Boolean, DateTime, ForeignKey, Index, Integer, JSON, String, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from .base import Base
@@ -81,7 +81,7 @@ class NodeRouteEntry(Base):
     local: Mapped[bool] = mapped_column(Boolean, default=False, nullable=False)
     primary: Mapped[bool] = mapped_column(Boolean, default=False, nullable=False)
 
-    origin_asn: Mapped[int | None] = mapped_column(Integer)
+    origin_asn: Mapped[int | None] = mapped_column(BigInteger)  # DN42 ASN 超 int32
     protocol: Mapped[str | None] = mapped_column(String(128))
     rpki: Mapped[str | None] = mapped_column(String(16))
     next_hop: Mapped[str | None] = mapped_column(String(128))
